Fix all warnings - Add `static_cast`s where necessary - Simplify the interface of the Text class (only deal with utf-8 encoded   `std::strings` and let the user manage the conversions) - Use c++20

+++ examples/magnifying_glass/magnifying_glass.cpp
@@ -56,14 +56,14 @@ int main() {
         window.setCamera(camera);
         const glm::vec2 win_size = window.getSize();
         window.setClipPlanes(0, win_size.x, win_size.y, 0);
-        window.setViewport(0, 0, win_size.x, win_size.y);
+        window.setViewport(0, 0, static_cast<int>(win_size.x), static_cast<int>(win_size.y));
 
         window.draw(rect);
         window.draw(frame);
 
         window.setCamera(frameCam);
-        window.setClipPlanes(zoomValue, frameSize - zoomValue, frameSize - zoomValue, zoomValue);
-        window.setViewport(framePos.x, win_size.y - framePos.y - frameSize, frameSize, frameSize);
+        window.setClipPlanes(static_cast<float>(zoomValue), static_cast<float>(frameSize - zoomValue), static_cast<float>(frameSize - zoomValue), static_cast<float>(zoomValue));
+        window.setViewport(framePos.x, static_cast<int>(win_size.y) - framePos.y - frameSize, frameSize, frameSize);
 
         window.draw(rect);
