+++ include/MyGL/Texture.hpp
@@ -113,7 +113,7 @@ namespace my
          * @param b The blue component
          * @param alpha The alpha component
          */
-        MYGL_EXPORT void setBorderColor(int r, int g, int b, int alpha = 255);
+        MYGL_EXPORT void setBorderColor(uint8_t r, uint8_t g, uint8_t b, uint8_t alpha = 255);
 
         /**
          * @brief Sets the border color used with GL_CLAMP_TO_BORDER wrapping method
